requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
atomicwrites==1.2.1
attrs==18.2.0
backports.functools-lru-cache==1.5
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
configparser==3.5.0
coverage==4.4.2
coveralls==1.5.0
cryptography==2.3.1
docopt==0.6.2
enum34==1.1.6
environs==4.0.0
flasgger==0.8.0
Flask==1.0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.13.0
funcsigs==1.0.2
futures==3.1.1
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
marshmallow==2.15.6
mccabe==0.6.1
mistune==0.8.3
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.7.1
psycopg2-binary==2.7.5
py==1.6.0
pycparser==2.18
PyJWT==1.4.2
pylint==1.9.3
pyOpenSSL==18.0.0
pytest==3.8.0
python-coveralls==2.9.1
python-dotenv==0.9.1
PyYAML==3.13
requests==2.19.1
scandir==1.9.0
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
